feat(6782): extract docs contract wave3 harness
