+++ crates/kamn-core/tests/docs_contract_wave3_harness/kolme_live_integration_docs.rs
@@ -1,0 +1,35 @@
+const DOC: &str = include_str!("../../../../docs/architecture/kolme-live-integration.md");
+
+#[test]
+fn doc_contains_process_isolation_marker_contracts() {
+    assert!(DOC.contains("transport_convergence_status"));
+    assert!(DOC.contains("libp2p_process_isolation_status"));
+    assert!(DOC.contains("libp2p_two_node_process_isolated_status"));
+    assert!(DOC.contains("libp2p_three_node_process_isolated_status"));
+    assert!(DOC.contains("local_heavy_runtime_budget_status"));
+    assert!(DOC.contains("elapsed_seconds"));
+    assert!(DOC.contains("max_seconds"));
+    assert!(DOC.contains("command_max_seconds"));
+    assert!(DOC.contains("runtime_provider_client_contract=KolmeRuntimeCommitLiveProvider"));
+}
+
+#[test]
+fn doc_contains_process_isolation_fail_closed_reasons() {
+    assert!(DOC.contains("local_full_stack_integration_policy_reason_taxonomy_version_mismatch"));
+    assert!(DOC
+        .contains("local_full_stack_integration_policy_libp2p_process_isolation_status_mismatch"));
+    assert!(DOC.contains(
+        "local_full_stack_integration_policy_libp2p_two_node_process_isolated_status_mismatch"
+    ));
+    assert!(DOC.contains(
+        "local_full_stack_integration_policy_libp2p_three_node_process_isolated_status_mismatch"
+    ));
+    assert!(DOC.contains(
+        "local_full_stack_integration_policy_libp2p_summary_three_node_partition_rejoin_status_mismatch"
+    ));
+    assert!(DOC.contains(
+        "local_full_stack_integration_policy_libp2p_summary_three_node_publish_drop_status_mismatch"
+    ));
+    assert!(DOC.contains("local_full_stack_integration_policy_runtime_budget_status_mismatch"));
+    assert!(DOC.contains("local_full_stack_integration_policy_runtime_budget_exceeded"));
+}
