+++ crates/kamn-core/tests/docs_contract_wave3_harness/token_config_docs.rs
@@ -1,0 +1,28 @@
+const DOC: &str = include_str!("../../../../docs/foundation/token-config.md");
+
+#[test]
+fn doc_contains_token_launch_handoff_evidence_contract() {
+    assert!(DOC.contains("## Token Launch Handoff Evidence Contract"));
+    assert!(DOC.contains("generate_token_launch_handoff_evidence_bundle.sh"));
+    assert!(DOC.contains("check_token_launch_handoff_policy.sh"));
+    assert!(DOC.contains("token_launch_handoff_contract_lane_contract.py"));
+    assert!(DOC.contains("run_token_launch_handoff_contract_lane.sh"));
+    assert!(DOC.contains("run_token_launch_handoff_deep_lane.sh"));
+    assert!(DOC.contains("fixtures/token_launch/handoff_invariant_cases.json"));
+}
+
+#[test]
+fn regression_requires_token_launch_handoff_guard_marker() {
+    // Regression: #714
+    assert!(DOC.contains(
+        "supply/allocation invariant drift and insufficient approvals force `NO-GO` (`Regression: #714`)."
+    ));
+}
+
+#[test]
+fn regression_requires_token_launch_shared_contract_marker() {
+    // Regression: #1270
+    assert!(DOC.contains(
+        "shared contract-lane module marker remains required for docs/contracts drift guard (`Regression: #1270`)."
+    ));
+}
